Fix interger overflow in _calc_iframe_target_size_one_pass_cbr
Fix same as in libvpx:
vp9:
https://chromium-review.googlesource.com/c/webm/libvpx/+/5545799
vp8:
https://chromium-review.googlesource.com/c/webm/libvpx/+/4793692

Bug: b/352414650
(cherry picked from commit e756a62b37a32abe4159b66a2380f546b6b95ea6)

+++ av1/encoder/ratectrl.c
@@ -2767,7 +2767,7 @@ int av1_calc_iframe_target_size_one_pass_cbr(const AV1_COMP *cpi) {
     if (rc->frames_since_key < framerate / 2) {
       kf_boost = (int)(kf_boost * rc->frames_since_key / (framerate / 2));
     }
-    target = ((16 + kf_boost) * rc->avg_frame_bandwidth) >> 4;
+    target = ((int64_t)(16 + kf_boost) * rc->avg_frame_bandwidth) >> 4;
   }
   return av1_rc_clamp_iframe_target_size(cpi, target);
 }
